refactor: Simplifies Learner namespace structure and removes unused methods.
