+++ src/Tutor.Core/LearnerModel/ILearnerRepository.cs
@@ -1,10 +1,7 @@
-﻿using Tutor.Core.LearnerModel.Learners;
-
-namespace Tutor.Core.LearnerModel
+﻿namespace Tutor.Core.LearnerModel
 {
     public interface ILearnerRepository
     {
-        Learner GetById(int learnerId);
         Learner GetByIndex(string index);
         Learner Save(Learner learner);
     }
