+++ src/Tutor.Core/LearnerModel/Learner.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Tutor.Core.DomainModel.KnowledgeComponents;
 using Tutor.Core.LearnerModel.Workspaces;
 
-namespace Tutor.Core.LearnerModel.Learners
+namespace Tutor.Core.LearnerModel
 {
     public class Learner
     {
